Updated for efficiency
Optimizations made to the generated SQL query, so that it targets more specific sets of data (this was pushed to production but left out of this repo.)

For any class that exposes a Dispose() function (such as Command and Connection), it's good practice to make sure that Dispose() is always called at the end of the method. If the scope of the object I instantiated is local, I should be using Using blocks for this.

Modified the generated SQL query (sSQL) to use parameterized values instead of pulling them straight from the string that's pulled into the method. The criteria for this method in particular will not usually come from user input, so there's less risk. This will still help protect it from garbage data and malformed queries.

Consider the use of Clone() for the sake of not affecting other threads.